Add test macro instead.
