+++ Cargo.toml
@@ -3,4 +3,5 @@ members = [
   "./pedersen",
   "./t256",
   "./t384",
+  "./tests",
 ]
