+++ t256/Cargo.toml
@@ -12,6 +12,7 @@ ark-r1cs-std = { version = "0.4.0", default-features = false, optional = true }
 ark-std = { version = "0.4.0", default-features = false }
 ark-secp256r1 = {default-features = false, git = "https://github.com/arkworks-rs/curves" }
 pedersen = { path="../pedersen" }
+pedersen-test-templates = { path="../tests"}
 rand = { version = "0.8.5" }
 rand_core = { version = "0.6.4" }
 merlin = { version = "3.0.0" }
