+++ t384/Cargo.toml
@@ -11,6 +11,7 @@ ark-ec = { version = "0.4.0", default-features = false }
 ark-r1cs-std = { version = "0.4.0", default-features = false, optional = true }
 ark-std = { version = "0.4.0", default-features = false }
 pedersen = { path = "../pedersen" }
+pedersen-test-templates = { path="../tests"}
 rand = { version = "0.8.5" }
 rand_core = { version = "0.6.4" }
 merlin = { version = "3.0.0" }
