+++ tests/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "pedersen-test-templates"
+version = "0.0.1"
+description = "A macro for testing Pedersen commitments"
+
+[dependencies]
+ark-ff = { version = "0.4.0", default-features = false }
+ark-ec = { version = "0.4.0", default-features = false }
+ark-r1cs-std = { version = "0.4.0", default-features = false, optional = true }
+ark-std = { version = "0.4.0", default-features = false }
+pedersen = { path="../pedersen" }
+rand = { version = "0.8.5" }
+rand_core = { version = "0.6.4" }
+merlin = { version = "3.0.0" }
+
