+++ tests/src/lib.rs
@@ -1,0 +1,2 @@
+#[macro_use]
+pub mod test_pedersen;
